+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # v1.2 (unrelased)
 
-* We now explicitly try `~/.cargo/bin/racer` if `racer isn't on path.
+* We now explicitly try `~/.cargo/bin/racer` if `racer` isn't on path.
+* We no longer offer completions inside comments by default (it tends
+  to be slow and rarely offers completions). See
+  `racer-complete-in-comments`.
 
 # v1.1